+++ Public/New-WiFiProfile.ps1
@@ -143,6 +143,9 @@ function New-WiFiProfile
     }
     finally
     {
-        Remove-WiFiHandle -ClientHandle $clientHandle
+        if ($clientHandle)
+        {
+            Remove-WiFiHandle -ClientHandle $clientHandle
+        }
     }
 }
